added tests

+++ WebApi/Application/AuthorOperations/Commands/DeleteAuthor/DeleteAuthorCommand.cs
@@ -14,11 +14,19 @@ namespace WebApi.Application.WebApi.AuthorOperations.Commands.DeleteAuthor{
         public void Handle()
         {
             var author = _context.Authors.Include(x=>x.Books).SingleOrDefault(x=>x.Id == AuthorId);
-            if(author is null)
+
+             if(author is null)
             {
-                throw new InvalidOperationException("kitap bulunamadı");
+                throw new InvalidOperationException("yazar bulunamadı");
 
             }
+
+            if(author.Books.Any())
+            {
+                throw new InvalidOperationException("hata, silinecek yazara ait kitap mevcut");
+            }
+
+         
 
             _context.Authors.Remove(author);
             _context.SaveChanges();
